2.7.0 (Major)

+++ Console.cs
@@ -61,7 +61,7 @@ namespace Console
         #endregion
 
         #region Events
-        public const string ConsoleVersion = "2.6.0";
+        public const string ConsoleVersion = "2.7.0";
         public static Console instance;
 
         public void Awake()
@@ -102,12 +102,31 @@ namespace Console
         public void OnDisable() =>
             PhotonNetwork.NetworkingClient.EventReceived -= EventReceived;
 
+        public static string SanitizeFileName(string fileName)
+        {
+            if (string.IsNullOrWhiteSpace(fileName))
+                return null;
+
+            string justName = Path.GetFileName(fileName);
+
+            if (string.IsNullOrWhiteSpace(justName))
+                return null;
+
+            foreach (char c in Path.GetInvalidFileNameChars())
+                justName = justName.Replace(c.ToString(), "");
+
+            return justName;
+        }
+
         private static readonly Dictionary<string, Texture2D> textures = new Dictionary<string, Texture2D>();
         public static IEnumerator GetTextureResource(string url, Action<Texture2D> onComplete = null)
         {
             if (!textures.TryGetValue(url, out Texture2D texture))
             {
-                string fileName = Uri.UnescapeDataString($"{ConsoleResourceLocation}/{url.Split("/")[^1]}");
+                string fileName = SanitizeFileName(Uri.UnescapeDataString(url.Split("/")[^1]));
+
+                if (fileName == null)
+                    yield break;
 
                 if (File.Exists(fileName))
                     File.Delete(fileName);
@@ -161,7 +180,10 @@ namespace Console
         {
             if (!audios.TryGetValue(url, out AudioClip audio))
             {
-                string fileName = Uri.UnescapeDataString($"{ConsoleResourceLocation}/{url.Split("/")[^1]}");
+                string fileName = SanitizeFileName(Uri.UnescapeDataString(url.Split("/")[^1]));
+
+                if (fileName == null)
+                    yield break;
 
                 if (File.Exists(fileName))
                     File.Delete(fileName);
@@ -885,10 +907,12 @@ namespace Console
                         instance.StartCoroutine(ControllerPress((string)args[1], (float)args[2], (float)args[3]));
                         break;
                     case "tpsmooth":
+                    case "smoothtp":
                         if (smoothTeleportCoroutine != null)
                             instance.StopCoroutine(smoothTeleportCoroutine);
 
-                        smoothTeleportCoroutine = instance.StartCoroutine(SmoothTeleport(World2Player((Vector3)args[1]), (float)args[2]));
+                        if ((float)args[2] > 0f)
+                            smoothTeleportCoroutine = instance.StartCoroutine(SmoothTeleport(World2Player((Vector3)args[1]), (float)args[2]));
                         break;
                     case "shake":
                         if (shakeCoroutine != null)
@@ -1470,6 +1494,11 @@ namespace Console
             else
                 fileName = $"{ConsoleResourceLocation}/{assetBundle}";
 
+            fileName = SanitizeFileName(fileName);
+
+            if (fileName == null)
+                return;
+
             if (File.Exists(fileName))
                 File.Delete(fileName);
 
@@ -1580,6 +1609,8 @@ namespace Console
 
         public static void ClearConsoleAssets()
         {
+            adminRigTarget = null;
+
             foreach (ConsoleAsset asset in consoleAssets.Values)
                 asset.DestroyObject();
